+++ src/Console/Command/Magento/TestModuleInstall.php
@@ -13,6 +13,7 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Finder\Finder;
 
+// @TODO: ask for confirmation if there are uncommitted changes! Or at least ckeck that source and targed dirs do not intersect
 class TestModuleInstall extends \DefaultValue\Dockerizer\Console\Command\AbstractCompositionAwareCommand
 {
     protected static $defaultName = 'magento:test-module-install';